requirements.txt
fastapi
pydantic
uvicorn
faiss-cpu
sentence-transformers
sqlite-utils
scikit-learn
flask
PyPDF2
rank_bm25
tiktoken
